Move maybe_fail_import() to top level test utils
